app/requirements.txt
fastapi
pydantic
asyncpg
bcrypt
python-dotenv
aiomysql
redis
pydantic_settings
PyJWT
fastapi-limiter==0.1.5 
cryptography
aiofiles
Pillow
python-multipart
requests
torch
transformers
pinecone==7.0.2
torchvision
typing
uvicorn[standard]